Moving the getCollections helper file into db/pg, improving it to match the standards of getAllCollections.js and removing getAllCollections as they were doing similar things
