+++ test/db/getCollections.test.js
@@ -1,14 +1,14 @@
 const tape = require('tape');
 const assert = require('assert');
 
-const redisCli = require('./configureRedis.js');
-const pg = require('./configurePool.js');
+const redisCli = require('../test_helpers/configureRedis.js');
+const pg = require('../test_helpers/configurePool.js');
 const pool = new pg.pool(pg.config);
 
 const flushDb = require('../helpers/flushDb.js')(pool, redisCli);
 const registerUser = require('../helpers/registerUser.js')(pool, redisCli);
-const getCollections = require('../helpers/getCollections.js')(pool);
-const createCollection = require('../helpers/createCollection.js')(pool);
+const getCollections = require('../../db/pg/getCollections.js')(pool);
+const createCollection = require('../../db/pg/createCollection.js')(pool);
 
 tape('getCollections', (t) => {
   const collectionObj = {
